Move top-level check logic to commands/check
